chore: update applications dependencies
