--- apps/play/global.d.ts
@@ -1,4 +1,0 @@
-declare module '*.md' {
-  const content: string;
-  export default content;
-}
