+++ apps/play/package.json
@@ -16,23 +16,20 @@
     "@arwes/styles": "^1.0.0-alpha.19",
     "@arwes/theme": "^1.0.0-alpha.19",
     "@arwes/tools": "^1.0.0-alpha.19",
-    "@mdi/js": "^6.6.95",
-    "@mdi/react": "^1.5.0",
-    "empanada": "^0.2.2",
-    "monaco-editor": "^0.33.0",
-    "noxtron": "^0.6.3"
+    "empanada": "^0.2.2"
   },
   "devDependencies": {
+    "@babel/core": "^7.18.2",
+    "@babel/preset-react": "^7.17.12",
+    "babel-loader": "^8.2.5",
     "copy-webpack-plugin": "^9.0.1",
-    "css-loader": "^6.7.1",
     "html-webpack-plugin": "^5.3.2",
-    "monaco-editor-webpack-plugin": "^7.0.1",
-    "raw-loader": "^4.0.2",
-    "style-loader": "^3.3.1"
+    "noxtron": "^0.7.1",
+    "raw-loader": "^4.0.2"
   },
   "scripts": {
     "build": "cross-env NODE_ENV=production webpack --progress",
     "dev": "webpack serve",
-    "start": "http-server ./build --port 9000 --gzip true -o /play"
+    "start": "serve ./build -p 9000"
   }
 }
