+++ apps/play/src/playground/playground.html
@@ -71,7 +71,7 @@
     <link rel="icon" href="https://next.arwes.dev/logo.png" type="image/png">
     <link rel="preconnect" href="https://fonts.gstatic.com" />
     <link rel="preconnect" href="https://fonts.googleapis.com" />
-    <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Source+Code+Pro&family=Titillium+Web:wght@400;700&display=swap">
+    <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Source+Code+Pro&family=Titillium+Web:wght@400;600&display=swap">
   </head>
   <body>
     <div id="root">
@@ -81,5 +81,6 @@
         alt="Arwes Playground"
       />
     </div>
+    <script src="/play/noxtron/playground.js"></script>
   </body>
 </html>
