--- apps/play/src/playground/types/empanada.md
@@ -1,7 +1,0 @@
-```ts
-declare module 'empanada' {
-  export const LOREM_IPSUM = string;
-  export const createRandomEmail: () => string;
-  export const createRandomWords: (length?: number) => string;
-}
-```
