+++ apps/play/src/sandbox/sandbox.html
@@ -25,5 +25,6 @@
   </head>
   <body>
     <div id="root"></div>
+    <script src="/play/noxtron/sandbox.js"></script>
   </body>
 </html>
