+++ package.json
@@ -44,7 +44,6 @@
     "eslint-plugin-react-hooks": "^4.3.0",
     "eslint-plugin-standard": "^4.1.0",
     "glob": "^7.2.0",
-    "http-server": "^14.1.0",
     "husky": "^7.0.4",
     "jest": "^27.5.1",
     "jest-canvas-mock": "^2.4.0",
@@ -53,6 +52,7 @@
     "markdownlint-cli": "^0.31.1",
     "npm-dts": "^1.3.11",
     "rimraf": "^3.0.2",
+    "serve": "^13.0.2",
     "ts-jest": "^27.1.3",
     "ts-loader": "^9.2.8",
     "tsconfig-paths-webpack-plugin": "^3.5.2",
